requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==2025.4.4
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.24
lxml==5.3.1
marshmallow==3.26.1
multidict==6.3.2
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
packaging==24.2
primp==0.14.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
zstandard==0.23.0
